+++ examples/resources/harness_platform_connector_terraform_cloud/import.sh
@@ -1,0 +1,8 @@
+# Import account level Terraform Cloud provider connector 
+terraform import harness_platform_connector_terraform_cloud.example <connector_id>
+
+# Import org level Terraform Cloud provider connector 
+terraform import harness_platform_connector_terraform_cloud.example <ord_id>/<connector_id>
+
+# Import project level Terraform Cloud provider connector 
+terraform import harness_platform_connector_terraform_cloud.example <org_id>/<project_id>/<connector_id>