+++ internal/service/platform/connector/aws_data_source.go
@@ -93,7 +93,7 @@ func DatasourceConnectorAws() *schema.Resource {
 		},
 	}
 
-	helpers.SetMultiLevelDatasourceSchema(resource.Schema)
+	helpers.SetMultiLevelDatasourceSchemaIdentifierRequired(resource.Schema)
 
 	return resource
 }