+++ internal/service/platform/connector/azure_cloud_provider_data_source.go
@@ -141,7 +141,7 @@ func DataSourceConnectorAzureCloudProvider() *schema.Resource {
 			},
 		},
 	}
-	helpers.SetMultiLevelDatasourceSchema(resource.Schema)
+	helpers.SetMultiLevelDatasourceSchemaIdentifierRequired(resource.Schema)
 
 	return resource
 }