+++ internal/service/platform/connector/nexus_data_source.go
@@ -58,7 +58,7 @@ func DatasourceConnectorNexus() *schema.Resource {
 		},
 	}
 
-	helpers.SetMultiLevelDatasourceSchema(resource.Schema)
+	helpers.SetMultiLevelDatasourceSchemaIdentifierRequired(resource.Schema)
 
 	return resource
 }