+++ internal/service/platform/connector/sumologic_data_source.go
@@ -35,7 +35,7 @@ func DatasourceConnectorSumologic() *schema.Resource {
 		},
 	}
 
-	helpers.SetMultiLevelDatasourceSchema(resource.Schema)
+	helpers.SetMultiLevelDatasourceSchemaIdentifierRequired(resource.Schema)
 
 	return resource
 }